// components/pdf/types.ts
export interface PDFStyles {
    page: any
    coverContainer: any
    coverTitle: any
    coverSubtitle: any
    coverInfo: any
    coverOrg: any
    header: any
    subheader: any
    categoryHeader: any
    table: any
    tableRow: any
    tableHeader: any
    cellBase: any
    image: any
    [key: string]: any
  }